Make single precision the default in VecCore
Since Vc doesn't support vectors of 64 bit types to match Int_v and
Real_v in the backend implementations with the same number of elements,
we shall default to single precision during the backend specification
and the creation of conformance tests. Once Vc adds support for vectors
of 64 bit integer types, the default for Real_v can be changed back to
double precision.

+++ include/VecCore/Backend/VcVector.h
@@ -11,7 +11,7 @@
 namespace vecCore {
 namespace backend {
 
-template <typename FloatT = double, bool EReturns = true>
+template <typename FloatT = float, bool EReturns = true>
 struct VcVector {
   static constexpr bool EarlyReturns = EReturns;
